Move `entity` module to `remote_control` app
`resolve` does not depend on the project startup.

After moving it, firstly, it makes our testing more convenient,
and secondly, it allows other modules of remote_control to reuse the resolve function of the entity.
